Update to vertx-auth split pkg remediation
