+++ src/main/java/io/vertx/ext/mail/impl/sasl/AuthOperationFactory.java
@@ -16,7 +16,7 @@
 
 package io.vertx.ext.mail.impl.sasl;
 
-import io.vertx.ext.auth.PRNG;
+import io.vertx.ext.auth.prng.PRNG;
 import io.vertx.ext.mail.MailConfig;
 import io.vertx.ext.mail.impl.Utils;
